[sw] Disable hello_usbdev verilator test
This has been really flaky in CI, and in general we have been trying to
keep examples separate from tests, so it seems reasonable to disable.

The reason not to is it's the only end-to-end test we have covering
usbdev at the moment. I'm not sure about how covered this module is with
other kinds of tests (including DV).

Signed-off-by: Sam Elliott <selliott@lowrisc.org>