[otbn] Auto-generate linker script based on otbn.hjson
This patch creates mem_layout.py, which uses reggen and the data in
otbn.hjson to find the memory layout for the block. We then use this
in otbn-ld to automatically make a linker script with the right
layout.

We'll also want to use it in the simulator (which needs to know the
size of memory) and when we load up ELF files in a block-level
testbench (which needs to know which LMAs correspond to imem and
dmem).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>