[otbn] Teach otbn-as some pseudo-operations
These all actually belong to RISC-V, so it was a bit difficult to
decide whether to "implement" them, or whether to just pass them
through to binutils.

For now, I decided to do the expansion ourselves. The advantage is
that then we can be absolutely certain no version of binutils will
start emitting instructions that OTBN doesn't support. The
disadvantage is that I might have got LI wrong.

So I've done some simple checks to make sure LI is right, assembling
the checked-in pseudo-ops.S with otbn-as and riscv32-unknown-elf-as
and checking the results were the same. (This exercises the 4
different generated sequences).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>