[otbn] default increment dmem pointer
for dmem access operations with two pointer increments
(reg pointer and dmem) pointer only one increment can be
handled in the load/store instruction. The other increment
is handeled in a subsequent `addi`. This adjusts the increments
such that always the dmem pointer is incremented in the load/store
and the reg pointer in the `addi`. This has the advantage that an
immediate of `1` can be used for the `addi` instruction regardless
of the type of addressing used for dmem.

Signed-off-by: Felix Miller <felix.miller@gi-de.com>