[CI] Check HW Tops in CI
Lots of chip information is generated by topgen, and committed to the
repo (into paths containing `autogen`).

There's a step in CI to check that the register definitions have not
changed. It uses `hw/Makefile` to do so. There have been updates to that
file to support running topgen, but no similar updates to the definition
in CI.

This change adds functionality to ensure topgen has been run and is
"clean" in CI.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>