[topgen] Add Auto-connect for Xbar
Previously, the connection between the module, memory list and the xbar
port shall be described manually in top["inter_module"]["connect"].

Now topgen, after generating the Xbar ips, adds the connections if
doesn't exist. It loops the node in every crossbars and check if it
comes from the module or connects to the other xbars. Then it adds
connection based on the information.

To make it easier and to not create double connection case,
`add_intermodule_connect()` function is added. The function checks if
the connection already exists and add if not.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>