[topgen] Remove unused options
The CI runs top_earlgrey_reg to check if any changes exist.
top_earlgrey_reg is to generate top_ral packages. It creates
intermediate hjson file.

This PR is to remove hjson-only option and no-gen-hjson option but only
checks top-ral to create ral pacakge and exit. So that it consistently
creates top_name.gen.hjson .

It also revises the dictionary variables to be OrderedDict
Signed-off-by: Eunchan Kim <eunchan@opentitan.org>