[rom_ext] Initial ROM_EXT Manifest Format
This commit contains a description of the initial ROM_EXT manifest
format. This documents every single field, and has tried to arrange them
in a reasonable order that satisfies any alignment requirements.

The version as described here can always change, this is good enough to
get started on an implementation of the code to produce a ROM_EXT ELF
executable.

Co-authored by: Silvestrs Timofejevs <silvestrst@lowrisc.org>

Signed-off-by: Sam Elliott <selliott@lowrisc.org>