[doc/security] Identities and Root Keys
The document describes the composition of the Silicon Creator and Silicon Owner
identities and the Silicon Owner root key derivation scheme. This scheme is based
on a symmetric key manager with support for software binding and key versioning.

This document also defines a non-cryptographic Device Identifier to facilitate
silicon tracking during manufacturing flows. The Device Identifier is also mixed
into the Creator Identity.

The document was converted to markdown by a GDoc plugin and formatted by
a md-linter. Since the maximum line width is set to 80cols, in most
cases a sentence change will trigger a full paragraph reformat/diff.

The GDoc to Markdown pluging generates tables in HTML format, which in
this case is useful to retain the per-cell text formatting and the rows
and columns layout.

Signed-off-by: Miguel Osorio <miguelosorio@google.com>