[regtool] Generate `NumAlerts` parameter from alert_list
With this commit, the regtool checks the number of alerts defined in
the alert_list defined in the hjson and creates a corresponding
parameter in the *_reg_pkg.sv. This parameter can then be used in the
RTL to e.g. instantiate the correct number of alert senders.

In addition, `AlertAsyncOn` is turned into a parameter. The reason for
this change is that ultimately, `AlertAsyncOn` will be driven by topgen
based on whether the peripheral and the alert handler are in the same
or different clock domains.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>