[meson] Put Toolchain Config into build-out
This commit moves the toolchain configuration file (which is created if
you use an old or non-standard toolchain) into `build-out` rather than
putting it into `/tmp`. It also ensures it has a timestamp that matches
that of the original `toolchain.txt` as committed to the repo. This
should prevent meson build loops, as noted in #2428

Signed-off-by: Sam Elliott <selliott@lowrisc.org>