[keccak] Add valid signal to random value
keccak_2share and its DOM AND operation requires 1600bits of random
value in each round. It currently takes 3 cycle in the masked keccak.
Preparing 1600 bits of entropy is hard to be done in 3 cycles.

There're a couple of options to prepare the entropy. One is reuse the
share of adjacent sheets, which is weaker than true entropy. Second is
to use smaller bits of entropy and expands it to 1600b. Another option
is to stack the entropy to become 1600b. To support the last scenario
(regardless of making true 1600b or smaller set then expanding to
1600b), now DOM has `c_valid_i` port. DOM latches internal registers
only when `c_valid_i` is high.

So outside of keccak_2share, the stimuli can indicate when the entropy
is good enough to be consumed.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>