[dv/rv_timer] fix regression corner case
There is a regression failure regarding timer timeout. The issue is
because the while loop is constructed :
```
          while (1) begin
            csr_rd(.ptr(intr_state_rg), .value(read_data));
            if (spinwait_delay_ns) #(spinwait_delay_ns * 1ns);
            if ((read_data == exp_data) | (reset_asserted == 1)) break;
          end
```
So in the worst case scenario, the interrupt is set right after csr_rd,
then the wait time should be twice of the spinwait_delay_ns.

This PR fixed this timeout issue by multiply the delay with 2.

Signed-off-by: Cindy Chen <chencindy@google.com>