[otbn] Automatically pass "-M numeric,no-aliases" in otbn-objdump
This stops the underlying disassembler from using RISC-V ABI names.
Since we aren't using the RISC-V ABI, we just want x0, x1 etc.

Similarly, we turn off aliases. While we do support li, nop and
ret (which are nice to get in the disassembly), we don't support a
whole host of other RISC-V pseudo-instructions (see tables 25.2 and
25.3 in the RISC-V Unprivileged ISA, version V20191213). It's probably
better not to disassemble to something we can't assemble again.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>