[docs] Address Builtin issues in C/C++ Style Guide
When trying to follow the style guide's requirements for builtins, it
was found that this was incompatible with some host library headers.

This change relaxes the C/C++ style guide rules on builtins, preserving
the notion of a "canonical definition" where an implementation can be
provided if a compiler does not provide a particular builtin or
intrinsic. The naming and declaration of this builtin has been relaxed
to account for C namespacing conventions.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>